Back to Git we come
mglover and I were working on a lot of behind-the-scenes code,
optimizing it for higher-resolution displays like MacBook Pro w/ retina
displays, and the iMac’s 5K display.  Unfortunately, we’ve also created
a lot of graphical bugs.

No worries, we’ll fix it soon.  We’re also open to fixes from you guys,
as the community.  If you have a fix, send a message to me with the
code and I’ll add it in ;)